Modify Wage, and Expense methods (working)

+++ src/main/java/org/ewallet/Expense.java
@@ -3,17 +3,17 @@ package org.ewallet;
 import java.util.Date;
 
 public class Expense {
-    private float amount;
+    private double amount;
     private Date date;
     private int yearlyFrequency; // 1 for 1 time or once a year, 12 for monthly or 24 for biweekly
 
-    public Expense(float amount, Date date, int yearlyFrequency) {
+    public Expense(double amount, Date date, int yearlyFrequency) {
         this.amount = amount;
         this.date = date;
         this.yearlyFrequency = yearlyFrequency;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
